+++ CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## 2.0 - 2019-08-02
+
+- Release the new API version 2.0, new functionalities
+  create a conversation
+  delete a conversation
+  create a new message attached to a conversation
+  link a message to another message in the same conversation
+  remove an existing link connection between two messages
+  delete a message from a conversation
+  add variables to a conversation
+
+ Deprecation
+`getMessage()` use getMessageByConversation() instead.
+
 ## 1.1 - 2018-02-01
-* Release the new API version 1.1.
-  ` getMessage()` is deprecated, please use getMessageByConversation() instead.
+
+- Release the new API version 1.1.
+  `getMessage()` is deprecated, please use getMessageByConversation() instead.